pyproject.toml
[project]
name = "hosvital-solutions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "colorama>=0.4.6",
    "pymssql>=2.3.2",
    "pyspark>=3.5.4",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "vaex>=4.17.0",
]
